# openapi/seed-api.yaml
openapi: 3.1.0
info:
  title: text-forge Seed API
  description: |
    前后端唯一契约事实源（规范第七章「契约先行」）。
    涵盖前端已使用的端点集合：项目 / 角色 / 生成 / 工作流。
    由 scripts/gen-api-types.mjs 生成 src/types/generated.ts（components.schemas.*）。
    后端实现须与该契约保持一致；前端调用处用 zod schema 校验响应以防脏数据白屏。
  version: 0.1.0
servers:
  - url: /api
    description: 相对路径（由前端 proxy / API_URL 解析）
...
paths:
  /projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: 获取项目列表
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    post:
      tags: [projects]
      operationId: createProject
      summary: 创建项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
...
  /characters:
    get:
      tags: [characters]
      operationId: listCharacters
      summary: 获取角色列表
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharactersResponse'
    post:
      tags: [characters]
      operationId: createCharacter
      summary: 创建角色
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
...
components:
...
  schemas:
    Project:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, generating, completed, paused]
        genre:
          type: string
        description:
          type: string
        pinned:
          type: boolean
        workflowId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectListResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    ProjectResponse:
      type: object
      required: [project]
      properties:
        project:
          $ref: '#/components/schemas/Project'
        version:
          type: integer
...
    CreateProjectRequest:
      type: object
      required: [title, description, genre]
      properties:
        title:
          type: string
        description:
          type: string
        genre:
          type: string
        version:
          type: integer
    Character:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        aliases:
          nullable: true
          type: array
          items:
            type: string
        description:
          type: string
        role:
          type: string
        status:
          type: string
        currentProfile:
          type: string
        customRole:
          type: string
        relationships:
          nullable: true
          type: array
          items:
            type: object
            required: [id, targetId, relation]
            properties:
              id:
                type: string
              targetId:
                type: string
              relation:
                type: string
        projectId:
          nullable: true
          type: string
        images:
          type: array
          items:
            type: string
        referenceImages:
          nullable: true
          type: array
          items:
            type: string
        referenceImage:
          nullable: true
          type: string
        imageSeed:
          nullable: true
          type: integer
        createdAt:
          type: string
    CharactersResponse:
      type: object
      required: [characters]
      properties:
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
    CharacterResponse:
      type: object
      required: [character]
      properties:
        character:
          $ref: '#/components/schemas/Character'
    CreateCharacterRequest:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        projectId:
          nullable: true
          type: string
        avatar:
          type: string
...